requirements.txt
Django==2.0.5
uWSGI==2.0.15
djangorestframework==3.7.7
markdown==2.6.11
django-filter==1.1.0
psycopg2==2.7.3.2
decorator==4.2.1
lxml==4.1.1
python-dateutil
factory-boy==2.10.0
phonenumbers==8.9.3
raven==6.1.0
git+https://github.com/wfmexpert/spyne.git
requests
django-compressor==2.2
XlsxWriter==1.0.5
social-auth-core[saml]==1.7.0 #libxmlsec1-dev
social-auth-app-django==2.1.0
python-ldap==3.1.0
xlwt==1.3.0
git+https://github.com/wfmexpert/verme-logs.git@1.24.1
django-axes==4.5.4
django-gfklookupwidget==1.0.6
transliterate
xlrd==1.1.0
django-admin-rangefilter==0.5.0

